Rename encoder utils file
